Move string module
