+++ neo/string.py
